readGain now only reads Gain, further functions should be easily writable

+++ Software/Firmware/smart-sensor/include/drivers/VEML7700.h
@@ -69,7 +69,7 @@ private:
 
     uint8_t testloop;
     uint16_t readConfig();
-    uint16_t readGain();
+    uint8_t readGain();
     uint8_t writeGain(uint8_t gain);
     uint8_t writeShutdown(uint8_t power);
 protected:
